style(storybook): 스토리북 폴더 구성 변경 및 Configure.mdx 추가
