+++ frontend/package.json
@@ -23,14 +23,14 @@
     "styled-components": "^6.0.7"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "^7.4.0",
-    "@storybook/addon-interactions": "^7.4.0",
-    "@storybook/addon-links": "^7.4.0",
+    "@storybook/addon-essentials": "^7.4.1",
+    "@storybook/addon-interactions": "^7.4.1",
+    "@storybook/addon-links": "^7.4.1",
     "@storybook/addon-onboarding": "^1.0.8",
-    "@storybook/addon-viewport": "^7.4.0",
-    "@storybook/blocks": "^7.4.0",
-    "@storybook/react": "^7.4.0",
-    "@storybook/react-vite": "^7.4.0",
+    "@storybook/addon-viewport": "^7.4.1",
+    "@storybook/blocks": "^7.4.1",
+    "@storybook/react": "^7.4.1",
+    "@storybook/react-vite": "^7.4.1",
     "@storybook/testing-library": "^0.2.0",
     "@testing-library/jest-dom": "^6.1.3",
     "@testing-library/react": "^14.0.0",
@@ -60,7 +60,8 @@
     "lint-staged": "^14.0.1",
     "msw": "^1.3.0",
     "prettier": "^3.0.3",
-    "storybook": "^7.4.0",
+    "storybook": "^7.4.1",
+    "storybook-addon-breakpoints": "^1.0.0",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.2",
