+++ frontend/src/components/layout/Footer/Footer.stories.tsx
@@ -1,13 +1,13 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Header from '../../../components/layout/Header';
+import Footer from '.';
 
 /**
- * 모먼트에서 사용하는 기본적인 헤더입니다.
+ * 모먼트에서 사용하는 기본적인 푸터입니다.
  */
 
-const meta: Meta<typeof Header> = {
-  title: 'Templates/Header',
-  component: Header,
+const meta: Meta<typeof Footer> = {
+  title: 'Templates/Footer',
+  component: Footer,
   tags: ['autodocs'],
   parameters: {
     layout: 'mobile',
@@ -22,8 +22,8 @@ const meta: Meta<typeof Header> = {
 };
 
 export default meta;
-type Story = StoryObj<typeof Header>;
+type Story = StoryObj<typeof Footer>;
 
 export const Default: Story = {
-  render: () => <Header />,
+  render: () => <Footer />,
 };
