+++ frontend/src/stories/Configure.mdx
@@ -1,0 +1,7 @@
+import { Meta } from '@storybook/blocks';
+
+<Meta title="Configure/Your project" />
+
+# Configure your project
+
+    Because Storybook works separately from your app, you'll need to configure it for your specific stack and setup. Below, explore guides for configuring Storybook with popular frameworks and tools. If you get stuck, learn how you can ask for help from our community.
